[add] utils.py
